很久后的一次提交

+++ src/main/java/com/zs/gms/controller/monitor/UnitController.java
@@ -12,6 +12,7 @@ import com.zs.gms.common.message.MessageEntry;
 import com.zs.gms.common.message.MessageFactory;
 import com.zs.gms.common.message.MessageResult;
 import com.zs.gms.common.service.GmsService;
+import com.zs.gms.common.utils.DateUtil;
 import com.zs.gms.common.utils.GmsUtil;
 import com.zs.gms.entity.mapmanager.SemiStatic;
 import com.zs.gms.entity.monitor.Unit;
@@ -58,32 +59,39 @@ public class UnitController extends BaseController {
     public void addUnit(@Valid @MultiRequestBody Unit unit, HttpServletResponse response) throws GmsException {
         try {
             User currentUser = super.getCurrentUser();
-            GmsResponse gmsResponse=null;
+            GmsResponse gmsResponse = null;
             if (!currentUser.getRoleSign().equals(Role.RoleSign.CHIEFDESPATCHER_ROLE.getValue())) {
                 gmsResponse = new GmsResponse().message("非调度长不能管理调度单元!").badRequest();
             }
             boolean existName = unitService.isExistName(unit.getUnitName());
             if (existName) {
-                gmsResponse =  new GmsResponse().message("调度单元名称已存在").badRequest();
+                gmsResponse = new GmsResponse().message("调度单元名称已存在").badRequest();
             }
-            User user = userService.findUserById(unit.getUserId());
-            if (user == null || !user.getRoleSign().equals(Role.RoleSign.DESPATCHER_ROLE.getValue())) {
-                gmsResponse =  new GmsResponse().message("选择的调度员不存在").badRequest();
+            Integer id = unit.getUserId();
+            if(null!=id){
+                User user = userService.findUserById(id);
+                if (user == null || !user.getRoleSign().equals(Role.RoleSign.DESPATCHER_ROLE.getValue())) {
+                    gmsResponse = new GmsResponse().message("选择的调度员不存在").badRequest();
+                }
             }
             Integer activeMap = MapDataUtil.getActiveMap();
             if (null == activeMap) {
-                gmsResponse =  new GmsResponse().message("没有处于活动中的地图!").badRequest();
+                gmsResponse = new GmsResponse().message("没有处于活动中的地图!").badRequest();
             }
             SemiStatic loadAreaInfo = MapDataUtil.getAreaInfo(activeMap, unit.getLoadAreaId());
             if (null == loadAreaInfo) {
-                gmsResponse =  new GmsResponse().message("选择的装载区不存在!").badRequest();
+                gmsResponse = new GmsResponse().message("选择的装载区不存在!").badRequest();
+            }
+            boolean existLoadId = unitService.isExistLoadId(unit.getLoadAreaId(),activeMap);
+            if (existLoadId) {
+                gmsResponse = new GmsResponse().message("该选定装载区已在其他调度单元中使用!").badRequest();
             }
             SemiStatic unloadAreaInfo = MapDataUtil.getAreaInfo(activeMap, unit.getUnLoadAreaId());
             if (null == unloadAreaInfo) {
-                gmsResponse =  new GmsResponse().message("选择的卸载区不存在!").badRequest();
+                gmsResponse = new GmsResponse().message("选择的卸载区不存在!").badRequest();
             }
-            if(null!=gmsResponse){
-                GmsService.callResponse(gmsResponse,response);
+            if (null != gmsResponse) {
+                GmsService.callResponse(gmsResponse, response);
                 return;
             }
             unit.setMapId(activeMap);
@@ -95,7 +103,7 @@ public class UnitController extends BaseController {
             paramMap.put("loaderAreaId", unit.getLoadAreaId());
             paramMap.put("unLoaderAreaId", unit.getUnLoadAreaId());
             paramMap.put("cycleTimes", unit.getCycleTimes());
-            paramMap.put("endTime", unit.getEndTime());
+            paramMap.put("endTime", GmsUtil.objNotNull(unit.getEndTime())? DateUtil.getDateFormat(unit.getEndTime(),DateUtil.FULL_TIME_SPLIT_PATTERN):"");
             MessageEntry entry = MessageFactory.createMessageEntry(GmsConstant.DISPATCH);
             entry.setAfterHandle(() -> {
                 if (!entry.getHandleResult().equals(MessageResult.SUCCESS)) {//操作失败
@@ -110,13 +118,18 @@ public class UnitController extends BaseController {
         }
     }
 
+    //tcl 0518 新增参数mapId
     @Log("获取调度单元列表")
     @GetMapping
     @ApiOperation(value = "获取调度单元列表", httpMethod = "GET")
     @ResponseBody
     public String getDispatchTaskList(QueryRequest queryRequest) throws GmsException {
         try {
-            Map<String, Object> dataTable = super.getDataTable(unitService.getUnitList(queryRequest));
+            Integer mapId = MapDataUtil.getActiveMap();
+            if(null==mapId){
+                return GmsUtil.toJsonIEnumDesc(new GmsResponse().message("活动地图不存在!").badRequest());
+            }
+            Map<String, Object> dataTable = super.getDataTable(unitService.getUnitList(queryRequest,mapId));
             return GmsUtil.toJsonIEnumDesc(new GmsResponse().data(dataTable).message("获取调度单元列表成功").success());
         } catch (Exception e) {
             String message = "获取调度单元列表失败";
@@ -154,17 +167,25 @@ public class UnitController extends BaseController {
     @Log("删除调度单元")
     @DeleteMapping("/{unitId}")
     @ApiOperation(value = "删除调度单元", httpMethod = "DELETE")
-    public GmsResponse deleteUnit(@PathVariable("unitId") Integer unitId) throws GmsException {
+    public void deleteUnit(@PathVariable("unitId") Integer unitId, HttpServletResponse response) throws GmsException {
         try {
             boolean existId = unitService.isExistId(unitId);
             if (!existId) {
-                return new GmsResponse().message("调度单元不存在").badRequest();
+                GmsService.callResponse(new GmsResponse().message("调度单元不存在").badRequest(), response);
+                return;
             }
-            unitService.deleteUnit(unitId);
-            unitVehicleService.clearVehiclesByUnitId(unitId);
-            return new GmsResponse().message("删除调度单元成功").success();
+            Map<String, Object> paramMap = new HashMap<>();
+            paramMap.put("unitId", unitId);
+            MessageEntry entry = MessageFactory.createMessageEntry(GmsConstant.DISPATCH);
+            entry.setAfterHandle(() -> {
+                if (entry.getHandleResult().equals(MessageResult.SUCCESS)) {//操作成功
+                    unitService.deleteUnit(unitId);
+                    unitVehicleService.clearVehiclesByUnitId(unitId);
+                }
+            });
+            MessageFactory.getDispatchMessage().sendMessageWithID(entry.getMessageId(), "RemoveAIUnit", JSONObject.toJSONString(paramMap), "取消调度单元成功");
         } catch (Exception e) {
-            String message = "删除调度单元失败";
+            String message = "取消调度单元失败";
             log.error(message, e);
             throw new GmsException(message);
         }
@@ -176,8 +197,12 @@ public class UnitController extends BaseController {
     @ResponseBody
     public String getUnits() throws GmsException {
         try {
+            Integer mapId = MapDataUtil.getActiveMap();
+            if(null==mapId){
+                return GmsUtil.toJsonIEnumDesc(new GmsResponse().message("活动地图不存在!").badRequest());
+            }
             User currentUser = this.getCurrentUser();
-            List<Unit> units = unitService.getUnitListByUserId(currentUser.getUserId());
+            List<Unit> units = unitService.getUnitListByUserId(currentUser.getUserId(),mapId);
             return GmsUtil.toJsonIEnumDesc(new GmsResponse().data(units).message("获取调度员管理的调度单元成功").success());
         } catch (Exception e) {
             String message = "获取调度员管理的调度单元失败";
